back/requirements.txt
openai
python-dotenv
requests
flask
flask-cors
google-genai
PyPDF2
python-docx
flask-socketio
eventlet